extension/icons/requirements.txt
Pillow
svglib
